+++ bundles/at.nucle.e4.plugin.preferences.core.api/src/at/nucle/e4/plugin/preferences/core/api/IPreferenceScope.java
@@ -1,0 +1,10 @@
+package at.nucle.e4.plugin.preferences.core.api;
+
+import org.eclipse.core.runtime.preferences.IScopeContext;
+
+public interface IPreferenceScope {
+
+	IScopeContext getScopeContext();
+
+	String getNode();
+}