+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 ## [Unreleased]
 
-## [0.1.0] - xxx-xx-xx
-Initial release
+## [0.2.0] - 2017-03-13
+
+### Added 
+- Interface IPreferenceScope
+- Define custom InstanceScope per bundle otherwise ist uses the default scope
+- PersistenceScopeRegistry
+- New extension point prefernce scope
+
+## [0.1.0] - 2017-03-12
 
 ### Added
-- Extension point for preference pages
+- Extension point for preference page
 - EPreferenceService as OSGI Service
 - Dependency Injection for preference page
 - Tycho build and folder structure
 - @Inject EPreferenceService pref;
 - Example with root-page and sub pages
+
+## [0.0.0] - 2017-03-11
+Initial release